Moved AddHeaderItem to QuoteRepository.cs changed conversion errors
